split off makeAutoUnitString
