+++ h5p/development/H5P.FormulaApplet-0.12/library.json
@@ -3,7 +3,7 @@
   "description": "Displays a Formula Applet",
   "majorVersion": 0,
   "minorVersion": 12,
-  "patchVersion": 135,
+  "patchVersion": 137,
   "runnable": 1,
   "author": "Rudolf Grossmann",
   "license": "MIT",
